TexturePack2Image/requirements.txt
Pillow==2.7.0
Pygments==2.0.1
Wand==0.3.9
bpython==0.13.1
colormath==2.1.0
ipython==2.3.1
